requirements.txt
# local package
-e .

# external requirements
fastapi==0.109.2
imbalanced_learn==0.12.0
joblib==1.3.2
matplotlib==3.8.2
numpy==1.26.4
pandas==2.2.0
PyYAML==6.0.1
scikit_learn==1.4.0
setuptools==69.0.3
uvicorn==0.27.1